+++ src/main/java/io/sniffy/Constants.java
@@ -28,6 +28,7 @@ public interface Constants {
      * The major version of Sniffy
      * @see Driver#getMinorVersion() ()
      */
-    int PATCH_VERSION = 3;
+    int PATCH_VERSION = 5;
+
 
 }