+++ src/test/java/io/sniffy/servlet/SnifferFilterTest.java
@@ -1,7 +1,6 @@
 package io.sniffy.servlet;
 
 import io.sniffy.BaseTest;
-import io.sniffy.BaseTest;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -13,8 +12,8 @@ import org.springframework.mock.web.MockServletContext;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
 import javax.servlet.*;
+import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletResponse;
-
 import java.io.IOException;
 import java.io.PrintWriter;
 
@@ -32,7 +31,6 @@ public class SnifferFilterTest extends BaseTest {
     private MockServletContext servletContext = new MockServletContext("/petclinic/");
     private MockHttpServletRequest httpServletRequest =
             MockMvcRequestBuilders.get("/petclinic/foo/bar?baz").contextPath("/petclinic").buildRequest(servletContext);
-    //new MockHttpServletRequest(servletContext, "GET", "/petclinic/foo/bar?baz");
     private SnifferFilter filter = new SnifferFilter();
 
     protected FilterConfig getFilterConfig() {
@@ -100,6 +98,26 @@ public class SnifferFilterTest extends BaseTest {
 
     }
 
+    @Test
+    public void testFilterThrowsException() throws IOException, ServletException {
+
+        FilterConfig filterConfig = getFilterConfig();
+        SnifferFilter filter = new SnifferFilter();
+        filter.init(filterConfig);
+
+        doAnswer(invocation -> {throw new RuntimeException("test");}).
+                when(filterChain).doFilter(any(), any());
+
+        try {
+            filter.doFilter(httpServletRequest, httpServletResponse, filterChain);
+            fail();
+        } catch (Exception e) {
+            assertNotNull(e);
+            assertEquals("test", e.getMessage());
+        }
+
+    }
+
 
     @Test
     public void testDisabledFilterOneQuery() throws IOException, ServletException {
@@ -114,6 +132,64 @@ public class SnifferFilterTest extends BaseTest {
 
         assertFalse(httpServletResponse.containsHeader(SnifferFilter.HEADER_NUMBER_OF_QUERIES));
 
+    }
+
+    @Test
+    public void testFilterEnabledByRequestParameter() throws IOException, ServletException {
+        doAnswer(invocation -> {executeStatement(); return null;}).
+                when(filterChain).doFilter(any(), any());
+        SnifferFilter filter = new SnifferFilter();
+        filter.setEnabled(false);
+        httpServletRequest.setParameter("sniffy", "true");
+        filter.doFilter(httpServletRequest, httpServletResponse, filterChain);
+        assertTrue(httpServletResponse.containsHeader(SnifferFilter.HEADER_NUMBER_OF_QUERIES));
+        assertEquals("Check cookie parameter specified", "true", httpServletResponse.getCookie("sniffy").getValue());
+    }
+
+    @Test
+    public void testFilterNoCookies() throws IOException, ServletException {
+        doAnswer(invocation -> {executeStatement(); return null;}).
+                when(filterChain).doFilter(any(), any());
+        SnifferFilter filter = new SnifferFilter();
+        filter.setEnabled(false);
+        httpServletRequest.setCookies(null);
+        filter.doFilter(httpServletRequest, httpServletResponse, filterChain);
+        assertFalse(httpServletResponse.containsHeader(SnifferFilter.HEADER_NUMBER_OF_QUERIES));
+    }
+
+    @Test
+    public void testFilterEnabledByCookie() throws IOException, ServletException {
+        doAnswer(invocation -> {executeStatement(); return null;}).
+                when(filterChain).doFilter(any(), any());
+        SnifferFilter filter = new SnifferFilter();
+        filter.setEnabled(false);
+        httpServletRequest.setCookies(new Cookie("sniffy", "true"));
+        filter.doFilter(httpServletRequest, httpServletResponse, filterChain);
+        assertTrue(httpServletResponse.containsHeader(SnifferFilter.HEADER_NUMBER_OF_QUERIES));
+    }
+
+    @Test
+    public void testFilterEnabledRequestParamOverridesCookie() throws IOException, ServletException {
+        doAnswer(invocation -> {executeStatement(); return null;}).
+                when(filterChain).doFilter(any(), any());
+        SnifferFilter filter = new SnifferFilter();
+        filter.setEnabled(false);
+        httpServletRequest.setParameter("sniffy", "false");
+        httpServletRequest.setCookies(new Cookie("sniffy", "true"));
+        filter.doFilter(httpServletRequest, httpServletResponse, filterChain);
+        assertFalse("Filter must be disabled", httpServletResponse.containsHeader(SnifferFilter.HEADER_NUMBER_OF_QUERIES));
+        assertEquals("Cookie parameter must be replaced", "false", httpServletResponse.getCookie("sniffy").getValue());
+    }
+
+    @Test
+    public void testFilterDisabledByRequestParameter() throws IOException, ServletException {
+        doAnswer(invocation -> {executeStatement(); return null;}).
+                when(filterChain).doFilter(any(), any());
+        SnifferFilter filter = new SnifferFilter();
+        filter.setEnabled(true);
+        httpServletRequest.setParameter("sniffy", "false");
+        filter.doFilter(httpServletRequest, httpServletResponse, filterChain);
+        assertFalse(httpServletResponse.containsHeader(SnifferFilter.HEADER_NUMBER_OF_QUERIES));
     }
 
     @Test
@@ -240,43 +316,6 @@ public class SnifferFilterTest extends BaseTest {
     }
 
     @Test
-    public void testDoNotInjectToXml() throws IOException, ServletException {
-
-        String actualContent = "<?xml version=\"1.0\"?>\n" +
-                "\n" +
-                "<applications location=\"PRODEXTSG\" hasPendingActions=\"false\">\n" +
-                "    <application>" +
-                "</application>\n" +
-                "    <featuredItems>\n" +
-                "        \n" +
-                "    </featuredItems>\n" +
-                "</applications>\n";
-
-        doAnswer(invocation -> {
-            HttpServletResponse response = (HttpServletResponse) invocation.getArguments()[1];
-
-            response.setContentType("text/html");
-
-            PrintWriter printWriter = response.getWriter();
-            executeStatement();
-            printWriter.append(actualContent);
-            executeStatement();
-            printWriter.flush();
-            return null;
-        }).when(filterChain).doFilter(any(), any());
-
-        SnifferFilter filter = new SnifferFilter();
-        filter.init(getFilterConfig());
-
-        filter.doFilter(httpServletRequest, httpServletResponse, filterChain);
-
-        assertEquals(2, httpServletResponse.getHeaderValue(SnifferFilter.HEADER_NUMBER_OF_QUERIES));
-        assertEquals(actualContent, httpServletResponse.getContentAsString());
-        assertFalse(httpServletResponse.getContentAsString().contains("id=\"sniffy\""));
-
-    }
-
-    @Test
     public void testInjectHtmlFlushResponse() throws IOException, ServletException {
 
         String actualContent = "<html><head><title>Title</title></head><body>Hello, World!</body></html>";
@@ -303,8 +342,40 @@ public class SnifferFilterTest extends BaseTest {
         assertEquals(1, httpServletResponse.getHeaderValue(SnifferFilter.HEADER_NUMBER_OF_QUERIES));
         String contentAsString = httpServletResponse.getContentAsString();
         assertTrue(contentAsString.contains("id=\"sniffy\""));
-        assertTrue(contentAsString.indexOf("id=\"sniffy\"") < contentAsString.indexOf("</body>"));
+        assertTrue(contentAsString.indexOf("id=\"sniffy\"") > contentAsString.indexOf("</body>"));
         assertTrue(contentAsString.contains("data-sql-queries=\"2\""));
+
+    }
+
+    @Test
+    public void testInjectHtmlCloseResponse() throws IOException, ServletException {
+
+        String actualContent = "<html><head><title>Title</title></head><body>Hello, World!</body></html>";
+
+        doAnswer(invocation -> {
+            HttpServletResponse response = (HttpServletResponse) invocation.getArguments()[1];
+
+            response.setContentType("text/html");
+
+            PrintWriter printWriter = response.getWriter();
+            executeStatement();
+            printWriter.append(actualContent);
+            printWriter.close();
+            executeStatement();
+
+            return null;
+        }).when(filterChain).doFilter(any(), any());
+
+        SnifferFilter filter = new SnifferFilter();
+        filter.init(getFilterConfig());
+
+        filter.doFilter(httpServletRequest, httpServletResponse, filterChain);
+
+        assertEquals(1, httpServletResponse.getHeaderValue(SnifferFilter.HEADER_NUMBER_OF_QUERIES));
+        String contentAsString = httpServletResponse.getContentAsString();
+        assertTrue(contentAsString.contains("id=\"sniffy\""));
+        assertTrue(contentAsString.indexOf("id=\"sniffy\"") < contentAsString.indexOf("</body>"));
+        assertTrue(contentAsString.contains("data-sql-queries=\"1\""));
 
     }